move EventSourcing.Persistence.PostgreSql DI code into main EventSourcing.Persistence.PostgreSql package
